requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=3.0.0
